Get train loss from update_parameters.
